we may have received a tcp FIN. in such case we should establish a new connection.
If the syslog server have closed the connection it will send a TCP FIN.
Current ensureSyslogServerConnection() implementation only check the socket state.
Those checks don't cover the TCP FIN case.
In order to check it, we should try and read from the socket.
If a FIN was sent, read() will return -1.
In such case we should mark the socket as invalid, and reconnect.
If we won't, the next message we'll send will be rejected with a RST.
No exception thrown outside to the caller, and no indication that the message was lost.